# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Covid19 App OpenApi Documentation",
        "description": "Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/covid": {
            "get": {
                "tags": [
                    "Covid"
                ],
                "summary": "Get continents covid info",
                "description": "Get Covid19 general info",
                "operationId": "getInfo",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/covid/getContinents": {
            "get": {
                "tags": [
                    "Covid"
                ],
                "summary": "Get continents covid info",
                "description": "",
                "operationId": "getContinents",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/covid/perContinent/{continent}": {
            "get": {
                "tags": [
                    "Covid"
                ],
                "summary": "Get countries info per continent",
                "description": "",
                "operationId": "getPerContinent",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/covid/perCountry/{country}": {
            "get": {
                "tags": [
                    "Covid"
                ],
                "summary": "Get info per country",
                "description": "",
                "operationId": "getByCountry",
                "parameters": [
                    {
                        "name": "country",
                        "in": "path",
                        "description": "Country name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Oauth2 security",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Covid",
            "description": "API Endpoints of Covid Info"
        }
    ]
}